Moved Question and QuestionManager to new package questions
